refactor: move and rename CollectorCurrentStatus
