+++ apps/web/src/services/asset-inventory/collector/shared/CollectorCurrentStatus.vue
